+++ src/main/java/spring/library/Service/MemberService.java
@@ -6,7 +6,6 @@ import org.springframework.stereotype.Service;
 import spring.library.DTO.MemberDto;
 import spring.library.DTO.MemberRequest;
 import spring.library.DTO.MemberResponse;
-import spring.library.DTO.MemberResponseNoId;
 import spring.library.Domain.Member;
 import spring.library.Repository.MemberRepository;
 
@@ -16,25 +15,37 @@ import java.util.List;
 @RequiredArgsConstructor
 public class MemberService {
 
+
     private final MemberRepository memberRepository;
 
-    public MemberResponseNoId addMember(MemberRequest memberRequest) {
-        return MemberResponseNoId.from(MemberDto.from(memberRepository.save(Member.from(memberRequest))));
+
+    public String addMember(MemberRequest memberRequest) {
+        memberRepository.save(Member.from(memberRequest));
+        return "추가 완료";
     }
+
 
     public List<MemberResponse> getAllMembers() {
         return memberRepository.findAll().stream().map(MemberDto::from).toList().stream().map(MemberResponse::from).toList();
     }
 
+
     @Transactional
-    public MemberResponseNoId updateMember(Long memberId, MemberRequest memberRequest) {
+    public String updateMember(Long memberId, MemberRequest memberRequest) {
         Member member = memberRepository.findById(memberId).orElse(null);
+        if(member == null) return "오류";
         member.update(memberRequest);
-        return MemberResponseNoId.from(MemberDto.from(member));
+        return "수정 완료";
     }
 
-    public void deleteMember(Long memberId) {
-        memberRepository.deleteById(memberId);
+
+    public String deleteMember(Long memberId) {
+        if(memberRepository.existsById(memberId)){
+            memberRepository.deleteById(memberId);
+            return "삭제 완료";
+        }
+        else return "오류";
     }
+
 
 }
